requirements.txt
aiopg==1.3.5
amqp==5.1.1
anyio==4.0.0
asgiref==3.5.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
billiard==4.1.0
cachetools==5.3.2
celery==5.3.1
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-postgres==0.0.5
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
dj-database-url==2.0.0
Django==4.1.13
django-redis==5.3.0
djangorestframework==3.14.0
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
gspread==5.12.0
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.1
httplib2==0.22.0
httpx==0.25.1
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
joblib==1.3.2
kombu==5.3.1
msgpack==1.0.7
ndg-httpsclient==0.5.1
nltk==3.8.1
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.8.0.76
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
priority==1.3.0
prompt-toolkit==3.0.39
protobuf==4.25.0
psycopg2-binary==2.9.7
pusher==3.3.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyfcm==1.5.4
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==5.0.0
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
timedelta==2020.12.3
tls-client==0.2.2
tqdm==4.66.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
zope.interface==6.1
